refactor: 💡 extract functions to helper file
