+++ services/viva-ms/lambdas/setCaseCompletions.js
@@ -5,25 +5,18 @@ import config from '../../../config';
 import * as dynamoDb from '../../../libs/dynamoDb';
 import params from '../../../libs/params';
 import log from '../../../libs/logs';
-import { getStatusByType } from '../../../libs/caseStatuses';
-import {
-  VIVA_COMPLETION_REQUIRED,
-  VIVA_RANDOM_CHECK_REQUIRED,
-  ACTIVE_COMPLETION_REQUIRED_VIVA,
-  ACTIVE_RANDOM_CHECK_REQUIRED_VIVA,
-} from '../../../libs/constants';
 
-import vivaAdapter from '../helpers/vivaAdapterRequestClient';
+import completionsHelper from '../helpers/completions';
 
 export async function main(event, context) {
   const { personalNumber } = event.detail.user;
 
   const [getLatestWorkflowIdError, latestWorkflowId] = await to(
-    getLatestVivaWorkflowId(personalNumber)
+    completionsHelper.get.workflow.latest(personalNumber)
   );
   if (getLatestWorkflowIdError) {
     log.error(
-      'Error getting Viva workflow completions',
+      'Error getting the latest Viva workflow',
       context.awsRequestId,
       'service-viva-ms-setCaseCompletions-001',
       getLatestWorkflowIdError
@@ -32,7 +25,7 @@ export async function main(event, context) {
   }
 
   const [getWorkflowCompletionsError, workflowCompletions] = await to(
-    getVivaWorkflowCompletions(personalNumber, latestWorkflowId)
+    completionsHelper.get.workflow.completions(personalNumber, latestWorkflowId)
   );
   if (getWorkflowCompletionsError) {
     log.error(
@@ -44,7 +37,9 @@ export async function main(event, context) {
     return false;
   }
 
-  const [getCaseError, userCase] = await to(getCase(personalNumber, latestWorkflowId));
+  const [getCaseError, userCase] = await to(
+    completionsHelper.get.caseOnWorkflowId(personalNumber, latestWorkflowId)
+  );
   if (getCaseError) {
     log.error(
       'Get case from cases table failed',
@@ -73,21 +68,19 @@ export async function main(event, context) {
     SK: userCase.SK,
   };
   const caseUpdateAttributes = {
-    newStatus: getCompletionStatus(workflowCompletions),
-    newState: getCompletionState(workflowCompletions),
-    newCurrentFormId: getCompletionFormId(vivaCaseSSMParams, workflowCompletions),
+    newStatus: completionsHelper.get.status(workflowCompletions),
+    newState: completionsHelper.get.state(workflowCompletions),
+    newCurrentFormId: completionsHelper.get.formId(vivaCaseSSMParams, workflowCompletions),
     newPersons: resetCasePersonsApplicantSignature(userCase),
     workflowCompletions,
   };
-  const [updateCaseCompletionsError, updatedCaseItem] = await to(
-    updateCase(caseKeys, caseUpdateAttributes)
-  );
-  if (updateCaseCompletionsError) {
+  const [updateCaseError, updatedCaseItem] = await to(updateCase(caseKeys, caseUpdateAttributes));
+  if (updateCaseError) {
     log.error(
       'Update case completion attributes failed',
       context.awsRequestId,
       'service-viva-ms-setCaseCompletions-005',
-      updateCaseCompletionsError
+      updateCaseError
     );
     return false;
   }
@@ -100,69 +93,6 @@ export async function main(event, context) {
   );
 
   return true;
-}
-
-function getCompletionFormId(caseSSMParams, completions) {
-  const { completionRandomCheckFormId, completionFormId } = caseSSMParams;
-  return completions.isRandomCheck ? completionRandomCheckFormId : completionFormId;
-}
-
-function getCompletionStatus(completions) {
-  return completions.isRandomCheck
-    ? getStatusByType(ACTIVE_RANDOM_CHECK_REQUIRED_VIVA)
-    : getStatusByType(ACTIVE_COMPLETION_REQUIRED_VIVA);
-}
-
-function getCompletionState(completions) {
-  return completions.isRandomCheck ? VIVA_RANDOM_CHECK_REQUIRED : VIVA_COMPLETION_REQUIRED;
-}
-
-async function getVivaWorkflowCompletions(personalNumber, workflowId) {
-  const [getWorkflowCompletionsError, getCompletionsResponse] = await to(
-    vivaAdapter.workflow.getCompletions({ personalNumber, workflowId })
-  );
-  if (getWorkflowCompletionsError) {
-    throw getWorkflowCompletionsError;
-  }
-
-  return getCompletionsResponse.attributes;
-}
-
-async function getLatestVivaWorkflowId(personalNumber) {
-  const [getLatestError, getLatestResponse] = await to(
-    vivaAdapter.workflow.getLatest(personalNumber)
-  );
-  if (getLatestError) {
-    throw getLatestError;
-  }
-
-  return getLatestResponse.attributes.workflowid;
-}
-
-async function getCase(personalNumber, workflowId) {
-  const PK = `USER#${personalNumber}`;
-
-  const queryParams = {
-    TableName: config.cases.tableName,
-    KeyConditionExpression: 'PK = :pk',
-    FilterExpression: 'details.workflowId = :workflowId',
-    ExpressionAttributeValues: {
-      ':pk': PK,
-      ':workflowId': workflowId,
-    },
-  };
-
-  const [queryError, queryResponse] = await to(dynamoDb.call('query', queryParams));
-  if (queryError) {
-    throw queryError;
-  }
-
-  const caseItem = queryResponse.Items[0];
-  if (!caseItem) {
-    throw `Case with workflow id: ${workflowId} not found`;
-  }
-
-  return caseItem;
 }
 
 async function updateCase(keys, caseUpdateAttributes) {
@@ -197,12 +127,7 @@ async function updateCase(keys, caseUpdateAttributes) {
 }
 
 function resetCasePersonsApplicantSignature(caseItem) {
-  const { persons } = caseItem;
-
-  if (persons == undefined) {
-    return [];
-  }
-
+  const { persons = [] } = caseItem;
   return persons.map(person => {
     if (person.role === 'applicant' && person.hasSigned) {
       person.hasSigned = false;
